requirements.txt
crewai==0.28.8
langchain-groq==0.1.2
langchain-community==0.0.31
streamlit
GitPython
PyGithub
